+++ internal/service/router/router.go
@@ -216,10 +216,14 @@ func (r *routerResource) Read(ctx context.Context, req resource.ReadRequest, res
 		UUID: data.ID.ValueString(),
 	})
 	if err != nil {
-		resp.Diagnostics.AddError(
-			"Unable to read router details",
-			utils.ErrorDiagnosticDetail(err),
-		)
+		if utils.IsNotFoundError(err) {
+			resp.State.RemoveResource(ctx)
+		} else {
+			resp.Diagnostics.AddError(
+				"Unable to read router details",
+				utils.ErrorDiagnosticDetail(err),
+			)
+		}
 		return
 	}